#22: Altered the ProcessCircleService, so anyone can create a new Circle, but only the Circle Administrator can make changes to a Circle.

+++ cws-core/src/main/java/io/javadog/cws/core/services/ProcessCircleService.java
@@ -7,6 +7,7 @@
  */
 package io.javadog.cws.core.services;
 
+import io.javadog.cws.api.common.Action;
 import io.javadog.cws.api.common.Constants;
 import io.javadog.cws.api.common.ReturnCode;
 import io.javadog.cws.api.common.TrustLevel;
@@ -49,41 +50,50 @@ public final class ProcessCircleService extends Serviceable<ProcessCircleRespons
     public ProcessCircleResponse perform(final ProcessCircleRequest request) {
         verifyRequest(request, Permission.PROCESS_CIRCLE);
         final ProcessCircleResponse response;
+        final Action action = request.getAction();
 
-        switch (request.getAction()) {
-            case CREATE:
-                response = createCircle(request);
-                break;
-            case UPDATE:
-                response = updateCircle(request);
-                break;
-            case DELETE:
-                response = deleteCircle(request);
-                break;
-            case ADD:
-                response = addTrustee(request);
-                break;
-            case ALTER:
-                response = alterTrustee(request);
-                break;
-            case REMOVE:
-                response = removeTrustee(request);
-                break;
-            default:
-                // Unreachable Code by design.
-                throw new CWSException(ReturnCode.ILLEGAL_ACTION, "Unsupported Action.");
+        if (action == Action.CREATE) {
+            // Anyone cay create a Circle, only a Circle Administrator of the
+            // Circle in question may perform other actions on Circles.
+            response = createCircle(request);
+        } else {
+            if (Objects.equals(Constants.ADMIN_ACCOUNT, member.getName()) || (trustees.get(0).getTrustLevel() == TrustLevel.ADMIN)) {
+                switch (request.getAction()) {
+                    case UPDATE:
+                        response = updateCircle(request);
+                        break;
+                    case DELETE:
+                        response = deleteCircle(request);
+                        break;
+                    case ADD:
+                        response = addTrustee(request);
+                        break;
+                    case ALTER:
+                        response = alterTrustee(request);
+                        break;
+                    case REMOVE:
+                        response = removeTrustee(request);
+                        break;
+                    default:
+                        // Unreachable Code by design.
+                        throw new CWSException(ReturnCode.ILLEGAL_ACTION, "Unsupported Action.");
+                }
+            } else {
+                response = new ProcessCircleResponse(ReturnCode.AUTHORIZATION_WARNING, "Only a Circle Administrator may perform this action.");
+            }
         }
 
         return response;
     }
 
     /**
-     * Creating a new Circle, can only be performed by the System Administrator,
-     * and requires the name of a Circle, and the initial Circle Administrator
-     * to be set. As part of creating the new Circle, a new Secret Key will be
-     * generated and added encrypted via the Circle Administrator. And a new
-     * folder (initial data record, used by all data belonging to the Circle),
-     * will also be added.
+     * <p>Creation of Circles can be made by the System Administrator provided
+     * that the System Administrator has set the Id of the initial Circle
+     * Administrator, which cannot be the System Administrator.</p>
+     *
+     * <p>Creation of Circles can also be made by other members, and if that
+     * is the case, then they will be set themselves as the initial Circle
+     * Administrator.</p>
      *
      * @param request Request Object with the Circle Name and Administrator
      * @return Response from the creation.
@@ -91,50 +101,64 @@ public final class ProcessCircleService extends Serviceable<ProcessCircleRespons
     private ProcessCircleResponse createCircle(final ProcessCircleRequest request) {
         final ProcessCircleResponse response;
 
-        if (Objects.equals(Constants.ADMIN_ACCOUNT, member.getName())) {
-            final String name = request.getCircleName();
-            final CircleEntity existing = dao.findCircleByName(name);
-
-            if (existing == null) {
-                final MemberEntity circleAdmin = dao.find(MemberEntity.class, request.getMemberId());
-
-                if (circleAdmin == null) {
+        final String name = request.getCircleName();
+        final CircleEntity existing = dao.findCircleByName(name);
+        if (existing != null) {
+            response = new ProcessCircleResponse(ReturnCode.IDENTIFICATION_WARNING, "A Circle with the requested name already exists.");
+        } else {
+            if (Objects.equals(Constants.ADMIN_ACCOUNT, member.getName())) {
+                // The administrator is creating a new Circle, which requires that
+                // a Member Id is provided as the new Circle Administrator.
+                final MemberEntity owner = dao.find(MemberEntity.class, request.getMemberId());
+                if (owner == null) {
                     response = new ProcessCircleResponse(ReturnCode.IDENTIFICATION_WARNING, "Cannot create a new Circle with a non-existing Circle Administrator.");
-                } else if (Objects.equals(Constants.ADMIN_ACCOUNT, circleAdmin.getName())) {
+                } else if (Objects.equals(owner.getName(), Constants.ADMIN_ACCOUNT)) {
                     response = new ProcessCircleResponse(ReturnCode.IDENTIFICATION_WARNING, "It is not allowed for the System Administrator to be part of a Circle.");
                 } else {
-                    final CircleEntity circle = new CircleEntity();
-                    circle.setName(name);
-                    dao.persist(circle);
-
-                    createRootFolder(circle);
-                    final KeyAlgorithm algorithm = settings.getSymmetricAlgorithm();
-                    final KeyEntity keyEntity = new KeyEntity();
-                    keyEntity.setAlgorithm(algorithm);
-                    keyEntity.setStatus(Status.ACTIVE);
-                    dao.persist(keyEntity);
-
-                    final SecretCWSKey key = crypto.generateSymmetricKey(keyEntity.getAlgorithm());
-                    final PublicKey publicKey = crypto.dearmoringPublicKey(circleAdmin.getPublicKey());
-                    final PublicCWSKey cwsPublicKey = new PublicCWSKey(circleAdmin.getAlgorithm(), publicKey);
-                    final String circleKey = crypto.encryptAndArmorCircleKey(cwsPublicKey, key);
-                    final TrusteeEntity trustee = new TrusteeEntity();
-                    trustee.setMember(circleAdmin);
-                    trustee.setCircle(circle);
-                    trustee.setKey(keyEntity);
-                    trustee.setTrustLevel(TrustLevel.ADMIN);
-                    trustee.setCircleKey(circleKey);
-                    dao.persist(trustee);
-
-                    response = new ProcessCircleResponse();
-                    response.setCircleId(circle.getExternalId());
+                    response = createCircle(owner, name);
                 }
             } else {
-                response = new ProcessCircleResponse(ReturnCode.IDENTIFICATION_WARNING, "A Circle with the requested name already exists.");
+                response = createCircle(member, name);
             }
-        } else {
-            response = new ProcessCircleResponse(ReturnCode.AUTHORIZATION_WARNING, "Only the System Administrator may create a new Circle.");
         }
+        return response;
+    }
+
+    /**
+     * <p>Creates the actual Circle with given Circle Name and Admin. The Circle
+     * will also be given a new encryption key and a default root folder for
+     * storing of all Data Objects.</p>
+     *
+     * @param circleAdmin The initial Circle Administrator
+     * @param name        The name of the new Circle
+     * @return Response Object with the new Circle Id
+     */
+    private ProcessCircleResponse createCircle(final MemberEntity circleAdmin, final String name) {
+        final CircleEntity circle = new CircleEntity();
+        circle.setName(name);
+        dao.persist(circle);
+
+        createRootFolder(circle);
+        final KeyAlgorithm algorithm = settings.getSymmetricAlgorithm();
+        final KeyEntity keyEntity = new KeyEntity();
+        keyEntity.setAlgorithm(algorithm);
+        keyEntity.setStatus(Status.ACTIVE);
+        dao.persist(keyEntity);
+
+        final SecretCWSKey key = crypto.generateSymmetricKey(keyEntity.getAlgorithm());
+        final PublicKey publicKey = crypto.dearmoringPublicKey(circleAdmin.getPublicKey());
+        final PublicCWSKey cwsPublicKey = new PublicCWSKey(circleAdmin.getAlgorithm(), publicKey);
+        final String circleKey = crypto.encryptAndArmorCircleKey(cwsPublicKey, key);
+        final TrusteeEntity trustee = new TrusteeEntity();
+        trustee.setMember(circleAdmin);
+        trustee.setCircle(circle);
+        trustee.setKey(keyEntity);
+        trustee.setTrustLevel(TrustLevel.ADMIN);
+        trustee.setCircleKey(circleKey);
+        dao.persist(trustee);
+
+        final ProcessCircleResponse response = new ProcessCircleResponse();
+        response.setCircleId(circle.getExternalId());
 
         return response;
     }
